+++ yql/essentials/tools/yql_complete/ya.make
@@ -20,6 +20,7 @@ PEERDIR(
 )
 
 SRCS(
+    completion_factory.cpp
     yql_complete.cpp
 )
